// src/app/models/kscard/kscard.model.ts
export interface KsCardModel {
    id?: number;
    code?: string;
    name?: string;
    select?: boolean;
    active?: number;
    addr1?: string;
    addr2?: string;
    capiblockCreadeddate?: string;
    capiblockNameCreatedby?: string;
    curratetype?: number;
    cyphcode?: string;
    specode?: number;
    explain?: string;
    recstatus?: number;
    siteid?: number;
}